move DiseaseAnnotationHelper class into core.helper paackage
